Fix reconcileMappings output
Now this function returns a map with the number of elements equal to
the number of replicas in CPMS. But we expect it to be equal to the
number of machines specified by the CPMS filter.

Sometimes desired replicas number may be different, so we should
use the real number instead.